% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polkadot.R
\name{binance_klines}
\alias{binance_klines}
\title{Get Binance Klines API data}
\usage{
binance_klines(start_time, interval = "1d", symbol = "DOTEUR")
}
\arguments{
\item{start_time}{\verb{[character]}\cr
Start time in UNIX date format with \code{origin = "1970-01-01"}.}

\item{interval}{\verb{[character]}\cr
The interval to get the data for. Defaults to one day.
Other possible values are '1m', '3m', '5m', '15m', '30m', '1h', '2h',
'4h', '6h', '8h', '12h', '1d', '3d', '1w', '1M'.}

\item{symbol}{\verb{[character]}\cr
The currency pair.}
}
\value{
\link[tibble:tibble]{tibble::tibble}
}
\description{
The time is returned in UNIX time format (milliseconds). To convert it to
a date format, do\if{html}{\out{<div class="sourceCode r">}}\preformatted{as.Date(as.POSIXct(<time> / 1e3 ,origin = "1970-01-01", tz = "GMT"))
}\if{html}{\out{</div>}}
}
\examples{
# 2021-06-24
binance_klines("1630808472")
}
\keyword{internal}
