% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cardano}
\alias{cardano}
\title{Example Cardano Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7 rows and 12 columns.
}
\usage{
cardano
}
\description{
Example Cardano CSV as retrieved from \url{https://pooltool.io} via the
"generic" CSV export
}
\keyword{internal}
