% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kusama.R
\name{parse_kusama}
\alias{parse_kusama}
\title{Parse Cardano Staking Rewards}
\usage{
parse_kusama(
  address,
  currency = "EUR",
  pp_lang = "EN",
  filename = "kusama.csv"
)
}
\value{

}
\description{
Retrieves staking rewards for a specific account from
\url{https://subscan.io}.
}
\examples{
parse_kusama("E4ZoURFwN9eHPxwwDaijY6zXnTvZ3AwFFQFsnSekRTW2sPQ")
}
