% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polkadot.R
\name{parse_polkadot}
\alias{parse_polkadot}
\title{Parse Cardano Staking Rewards}
\usage{
parse_polkadot(
  address,
  currency = "EUR",
  pp_lang = "EN",
  filename = "polkadot.csv"
)
}
\value{

}
\description{
Retrieves staking rewards for a specific account from \url{https://subscan.io}.
}
\examples{
parse_polkadot("1qEk2g6N1uugFgyvmnsL6P9Conh5nTwL5mj85bm3XHtjc73")
}
