% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solana.R
\name{request_fun_sol}
\alias{request_fun_sol}
\title{Make API requests to Polkadot network}
\usage{
request_fun_sol(address, cursor = NULL, api_key)
}
\arguments{
\item{address}{\verb{[character]}\cr
Wallet address.}

\item{api_key}{\link{character}\cr
Optional personal API key.}

\item{page}{\verb{[integer]}\cr
The page to query.}

\item{row}{\verb{[integer]}\cr
How many rows to query per page.}
}
\value{
httr2 response
}
\description{
Make API requests to Polkadot network
}
\keyword{internal}
