% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardano.R
\name{parse_cardano}
\alias{parse_cardano}
\title{Parse Cardano Staking Rewards}
\usage{
parse_cardano(data, pp_lang = "DE", filename = "cardano.csv")
}
\arguments{
\item{data}{\link{character}\cr
The (generic) input CSV from pooltool.io.}

\item{pp_lang}{\link{character}\cr
The currency of Portfolio Performance. Only German ("DE") and English
("EN") are supported.}

\item{filename}{\link{character}\cr
The output filename.}

\item{currency}{}
}
\value{
\link{data.frame} (Invisible)
}
\description{
Reads generic CSV outputs from pooltool.io and transposes these into a format
which can be read by Portfolio Performance.

Rewards are classified as "Delivery (Inbound)" (German: "Einlieferung") which
is a category that is different from "buy" and "interest".
}
\examples{
data("cardano")
parse_cardano(cardano, pp_lang = "DE")
parse_cardano(cardano, pp_lang = "EN")
}
