% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardano.R
\name{parse_cardano}
\alias{parse_cardano}
\title{Parse Cardano Staking Rewards}
\usage{
parse_cardano(
  data,
  pp_security_name = "Cardano",
  pp_lang = "DE",
  securities_account = NULL,
  filename = NULL,
  dec = NULL
)
}
\arguments{
\item{data}{\link{character}\cr
The (generic) input CSV from \url{https://pooltool.io}.}

\item{pp_security_name}{\link{character}\cr
The name of the security in Portfolio Performance (e.g. "Kusama").}

\item{pp_lang}{\link{character}\cr
The currency of Portfolio Performance. Only German ("DE") and English
("EN") are supported.}

\item{securities_account}{\link{character}\cr
The name of the account to which to associate the deliveries.
In German called "Depot".}

\item{filename}{\link{character}\cr
The output filename.}

\item{dec}{\link{character} \cr
Decimal separator. If not set, inferred by the value of argument \code{pp_lang}:
\itemize{
\item If \code{pp_lang = "DE"} -> \code{dec = ","}
\item If \code{pp_lang = "EN"} -> \code{dec = "."}
}}
}
\value{
\link{data.frame} (Invisible)
}
\description{
Reads generic CSV outputs from pooltool.io and transposes these into a format
which can be read by Portfolio Performance.

Rewards are classified as "Delivery (Inbound)" (German: "Einlieferung") which
is a category that is different from "buy" and "interest".
}
\examples{
data("cardano")
parse_cardano(cardano, pp_lang = "DE")
parse_cardano(cardano, pp_lang = "EN")
}
