% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polkadot.R
\name{parse_polkadot}
\alias{parse_polkadot}
\title{Parse Polkadot Staking Rewards}
\usage{
parse_polkadot(
  address,
  pp_security_name = "Polkadot",
  currency = "EUR",
  pp_lang = "EN",
  securities_account = NULL,
  dec = NULL,
  filename = NULL,
  api_key = NULL
)
}
\arguments{
\item{address}{\verb{[character]}\cr
Wallet address.}

\item{pp_security_name}{\link{character}\cr
The name of the security in Portfolio Performance (e.g. "Kusama").}

\item{currency}{\verb{[character]}\cr
Currency in which to calculate the staking rewards.
Must be part of a valid coin pair on Binance, e.g. "DOTEUR".}

\item{pp_lang}{\link{character}\cr
The currency of Portfolio Performance. Only German ("DE") and English
("EN") are supported.}

\item{securities_account}{\link{character}\cr
The name of the account to which to associate the deliveries.
In German called "Depot".}

\item{dec}{\verb{[character]} \cr
Decimal separator. If not set, inferred by the value of argument \code{pp_lang}:
\itemize{
\item If \code{pp_lang = "DE"} -> \code{dec = ","}
\item If \code{pp_lang = "EN"} -> \code{dec = "."}
}}

\item{filename}{\link{character}\cr
The output filename.}

\item{api_key}{\link{character}\cr
Optional personal API key.}
}
\value{
\link[tibble:tibble]{tibble::tibble} (invisibly)
}
\description{
Retrieves staking rewards for a specific account from \url{https://subscan.io}.
}
\examples{
parse_polkadot("1qEk2g6N1uugFgyvmnsL6P9Conh5nTwL5mj85bm3XHtjc73")
parse_polkadot("1qEk2g6N1uugFgyvmnsL6P9Conh5nTwL5mj85bm3XHtjc73",
  api_key = "836063806b8c9dd642167210a05aa201"
)
}
